+++ pkg/services/secrets/manager/helpers_test.go
@@ -1,4 +1,4 @@
-package secretshelper
+package manager
 
 import (
 	"testing"
@@ -7,15 +7,16 @@ import (
 
 	"github.com/grafana/grafana/pkg/services/secrets/database"
 
+	"github.com/stretchr/testify/require"
+	"gopkg.in/ini.v1"
+
 	"github.com/grafana/grafana/pkg/bus"
 	"github.com/grafana/grafana/pkg/services/encryption/ossencryption"
 	"github.com/grafana/grafana/pkg/services/sqlstore"
 	"github.com/grafana/grafana/pkg/setting"
-	"github.com/stretchr/testify/require"
-	"gopkg.in/ini.v1"
 )
 
-func SetupTestService(t *testing.T) secrets.SecretsService {
+func SetupTestService(t *testing.T) (secrets.SecretsService, secrets.Store) {
 	t.Helper()
 	defaultKey := "SdlklWklckeLS"
 	if len(setting.SecretKey) > 0 {
@@ -27,10 +28,12 @@ func SetupTestService(t *testing.T) secrets.SecretsService {
 	require.NoError(t, err)
 	settings := &setting.OSSImpl{Cfg: &setting.Cfg{Raw: raw}}
 
-	return secrets.ProvideSecretsService(
-		database.ProvideSecretsStore(sqlstore.InitTestDB(t)),
+	store := database.ProvideSecretsStore(sqlstore.InitTestDB(t))
+
+	return ProvideSecretsService(
+		store,
 		bus.New(),
 		ossencryption.ProvideService(),
 		settings,
-	)
+	), store
 }
